requirements.txt
aiohttp==3.8.4
alabaster==0.7.13
async-timeout==4.0.2
asyncio==3.4.3
atomicwrites==1.4.1
attrs==22.2.0
Babel==2.12.1
bleach==6.0.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
discord-py-slash-command==4.2.1
discord.py==2.2.2
docutils==0.19
dropbox==11.36.0
ffmpeg==1.4
Flask==2.2.3
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
keyring==23.13.1
MarkupSafe==2.1.2
multidict==6.0.4
oauthlib==3.2.2
packaging==23.0
pathlib==1.0.1
pkginfo==1.9.6
pluggy==1.0.0
ply==3.11
protobuf==4.22.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
PyNaCl==1.5.0
pyparsing==3.0.9
pytest==7.2.2
pytest-runner==6.0.0
python-dotenv==1.0.0
pytz==2022.7.1
pywin32-ctypes==0.2.0
readme-renderer==37.3
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3986==2.0.0
rsa==4.9
selenium==4.8.2
six==1.16.0
snowballstemmer==2.2.0
Sphinx==6.1.3
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stone==3.3.1
toml==0.10.2
tqdm==4.65.0
twine==4.0.2
typing-extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
webencodings==0.5.1
Werkzeug==2.2.3
yarl==1.8.2
youtube-dl==2021.12.17
zipp==3.15.0
cryptography==39.0.2
fakeyou==1.2.4
